Added Hat Functionality
Player Marble can now give itself a hat based on the _hatId given to it upon its instantiation. For now, as the Player is already in the scene and not instantiated by a Scene Controller, I added a variable in the inspector that allows you to set what hat the Player should have.

In the future, another controller may need to handle the hatIds instead, and the Player's InstantiateHat() method may need to made public so the controller can access it instead.