Added Fancy Hat Models
All hats (minus the Ushanka) work perfectly with the simple instantiation script already in the code. Meaning we don't have to have specific transform values for each one. For the Ushanka though, we can either make a special case for it or just not use it. Adding a special case shouldn't be hard.