Added Scoring Method
Scene Controller now has scoring method. Call CalculateScore() and pass in a list of player marbles. It will return a score based on the distances between the player marbles and the scoring marble. Also included is an out of bounds sprite prefab. Currently there is an instance of it under the scoring marble in the main scene.