Implemented Scoring Calculations / Fixed Hat Bug
Scene controller now gets both players' scores at the end of the game. Scores are saved as score1 and score2.

Hats now will only follow the rotation of a marble if the marble it is a child of is tagged "demoMarble". This prevents the hats from rotating sporadically during gameplay.

Also, scoring marble has be better centered to the board.